+++ Tests/XcodeGenKitTests/SpecLoadingTests.swift
@@ -147,12 +147,14 @@ class SpecLoadingTests: XCTestCase {
                     ["target": "name", "embed": false],
                     ["carthage": "name"],
                     ["framework": "path", "weak": true],
+                    ["sdk": "Contacts.framework"],
                 ]
                 let target = try Target(name: "test", jsonDictionary: targetDictionary)
-                try expect(target.dependencies.count) == 3
+                try expect(target.dependencies.count) == 4
                 try expect(target.dependencies[0]) == Dependency(type: .target, reference: "name", embed: false)
                 try expect(target.dependencies[1]) == Dependency(type: .carthage, reference: "name")
                 try expect(target.dependencies[2]) == Dependency(type: .framework, reference: "path", weakLink: true)
+                try expect(target.dependencies[3]) == Dependency(type: .sdk, reference: "Contacts.framework")
             }
 
             $0.it("parses info plist") {
@@ -161,15 +163,15 @@ class SpecLoadingTests: XCTestCase {
                     "path": "Info.plist",
                     "properties": [
                         "CFBundleName": "MyAppName",
-                        "UIBackgroundModes": ["fetch"]
-                    ]
+                        "UIBackgroundModes": ["fetch"],
+                    ],
                 ]
 
                 let target = try Target(name: "", jsonDictionary: targetDictionary)
                 try expect(target.info) == Plist(path: "Info.plist", attributes: [
                     "CFBundleName": "MyAppName",
-                    "UIBackgroundModes": ["fetch"]
-                    ])
+                    "UIBackgroundModes": ["fetch"],
+                ])
             }
 
             $0.it("parses entitlement plist") {
@@ -178,13 +180,13 @@ class SpecLoadingTests: XCTestCase {
                     "path": "app.entitlements",
                     "properties": [
                         "com.apple.security.application-groups": "com.group",
-                    ]
+                    ],
                 ]
 
                 let target = try Target(name: "", jsonDictionary: targetDictionary)
                 try expect(target.entitlements) == Plist(path: "app.entitlements", attributes: [
                     "com.apple.security.application-groups": "com.group",
-                    ])
+                ])
             }
 
             $0.it("parses cross platform targets") {
@@ -253,7 +255,7 @@ class SpecLoadingTests: XCTestCase {
             $0.it("parses target schemes") {
                 var targetDictionary = validTarget
                 targetDictionary["scheme"] = [
-                    "testTargets": ["t1", "t2"],
+                    "testTargets": ["t1", ["name": "t2"]],
                     "configVariants": ["dev", "app-store"],
                     "commandLineArguments": [
                         "ENV1": true,
@@ -312,6 +314,18 @@ class SpecLoadingTests: XCTestCase {
                             ],
                         ],
                     ],
+                    "test": [
+                        "config": "debug",
+                        "targets": [
+                            "Target1",
+                            [
+                                "name": "Target2",
+                                "parallelizable": true,
+                                "randomExecutionOrder": true,
+                            ],
+                        ],
+                        "gatherCoverageData": true,
+                    ],
                 ]
                 let scheme = try Scheme(name: "Scheme", jsonDictionary: schemeDictionary)
                 let expectedTargets: [Scheme.BuildTarget] = [
@@ -330,6 +344,20 @@ class SpecLoadingTests: XCTestCase {
 
                 try expect(scheme.build.parallelizeBuild) == false
                 try expect(scheme.build.buildImplicitDependencies) == false
+
+                let expectedTest = Scheme.Test(
+                    config: "debug",
+                    gatherCoverageData: true,
+                    targets: [
+                        "Target1",
+                        Scheme.Test.TestTarget(
+                            name: "Target2",
+                            randomExecutionOrder: true,
+                            parallelizable: true
+                        ),
+                    ]
+                )
+                try expect(scheme.test) == expectedTest
             }
 
             $0.it("parses schemes variables") {